requirements.txt
python-dotenv
pymongo
python-telegram-bot
google-ai-generativelanguage
google-generativeai
requests
schedule
python-dateutil
pandas